Adding example .env.example file to showcase variables; renaming authDefinitions.ts to kebap case
