+++ apps/endatix-hub/.env.example
@@ -1,0 +1,10 @@
+# Required Env Variables
+############################
+
+# The url where the Endatix API is hosted, e.g. https://api.endatix.com.
+# 💡 TIP: Default value during development is https://localhost:5001
+ENDATIX_BASE_URL=https://localhost:5001
+
+# The secrete secret key that will be used to encrypt the Endatix Hub session cookie. Example: a28110cc8b94c5f3b3c923aa2c9fae4ed50c86ec61debfe6edb3c29947dbb00c
+# 💡 TIP: You can use: `openssl rand -hex 32` to generate one
+SESSION_SECRET=DbLkxcw0C5om6n2MFdn6E4sevnT8RzWqcdhQ1FUzhiQ6eGm
